configure View controller
