--- VIP-Clean-Architecture/Scene/DataConsumption/DataConsumptionController.swift
@@ -1,33 +1,0 @@
-//
-//  ViewController.swift
-//  VIP-Clean-Architecture
-//
-//  Created by Ahil Ahilendran on 5/8/20.
-//  Copyright © 2020 Ahil Ahilendran. All rights reserved.
-//
-
-import UIKit
-
-class DataConsumptionController: UIViewController {
-    
-    let containerView = UIView()
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        setupViews()
-    }
-    
-    func setupViews() {
-        
-        containerView.backgroundColor = #colorLiteral(red: 0.9529411793, green: 0.6862745285, blue: 0.1333333403, alpha: 1)
-        view.addSubview(containerView)
-        
-        containerView.translatesAutoresizingMaskIntoConstraints = false
-        containerView.topAnchor.constraint(equalTo: view.topAnchor).isActive = true
-        containerView.leadingAnchor.constraint(equalTo: view.leadingAnchor).isActive = true
-        containerView.bottomAnchor.constraint(equalTo: view.bottomAnchor).isActive = true
-        containerView.rightAnchor.constraint(equalTo: view.rightAnchor).isActive = true
-    
-    }
-}
-
