+++ VIP-Clean-Architecture/Scene/DataConsumption/DataConsumptionInteractor.swift
@@ -21,7 +21,6 @@ final class DataConsumptionInteractor {
     }
 }
 
-// MARK: - PPHomeInteractorInput
 extension DataConsumptionInteractor: DataConsumptionInteractorInput {
     func getMobileDataConsumption() {
         apiWorker.getDataConsumption(offset: 10, resourceId: "a807b7ab-6cad-4aa6-87d0-e283a7353a0f", limit: 10)
